move workingTime stuff into workingtime package
