+++ src/main/java/de/focusshift/zeiterfassung/report/ReportServiceRaw.java
@@ -12,8 +12,8 @@ import de.focusshift.zeiterfassung.user.UserIdComposite;
 import de.focusshift.zeiterfassung.usermanagement.User;
 import de.focusshift.zeiterfassung.usermanagement.UserLocalId;
 import de.focusshift.zeiterfassung.usermanagement.UserManagementService;
-import de.focusshift.zeiterfassung.usermanagement.WorkingTimeCalendar;
-import de.focusshift.zeiterfassung.usermanagement.WorkingTimeCalendarService;
+import de.focusshift.zeiterfassung.workingtime.WorkingTimeCalendar;
+import de.focusshift.zeiterfassung.workingtime.WorkingTimeCalendarService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
