+++ src/main/java/de/focusshift/zeiterfassung/usermanagement/WorkingTimeController.java
@@ -2,6 +2,11 @@ package de.focusshift.zeiterfassung.usermanagement;
 
 import de.focus_shift.launchpad.api.HasLaunchpad;
 import de.focusshift.zeiterfassung.timeclock.HasTimeClock;
+import de.focusshift.zeiterfassung.workingtime.WorkDay;
+import de.focusshift.zeiterfassung.workingtime.WorkWeekUpdate;
+import de.focusshift.zeiterfassung.workingtime.WorkingTime;
+import de.focusshift.zeiterfassung.workingtime.WorkingTimeId;
+import de.focusshift.zeiterfassung.workingtime.WorkingTimeService;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.oauth2.core.oidc.user.OidcUser;
@@ -33,7 +38,7 @@ import java.util.function.Supplier;
 import static de.focusshift.zeiterfassung.security.SecurityRole.ZEITERFASSUNG_OVERTIME_ACCOUNT_EDIT_ALL;
 import static de.focusshift.zeiterfassung.security.SecurityRole.ZEITERFASSUNG_WORKING_TIME_EDIT_ALL;
 import static de.focusshift.zeiterfassung.usermanagement.UserManagementController.hasAuthority;
-import static de.focusshift.zeiterfassung.usermanagement.WorkingTime.hoursToDuration;
+import static de.focusshift.zeiterfassung.workingtime.WorkingTime.hoursToDuration;
 import static java.time.DayOfWeek.FRIDAY;
 import static java.time.DayOfWeek.MONDAY;
 import static java.time.DayOfWeek.SATURDAY;
