+++ src/main/java/de/focusshift/zeiterfassung/workingtime/WorkDay.java
@@ -1,4 +1,4 @@
-package de.focusshift.zeiterfassung.usermanagement;
+package de.focusshift.zeiterfassung.workingtime;
 
 import java.math.BigDecimal;
 import java.time.DayOfWeek;
