+++ src/main/java/de/focusshift/zeiterfassung/workingtime/WorkTimeServiceImpl.java
@@ -1,6 +1,9 @@
-package de.focusshift.zeiterfassung.usermanagement;
+package de.focusshift.zeiterfassung.workingtime;
 
 import de.focusshift.zeiterfassung.user.UserIdComposite;
+import de.focusshift.zeiterfassung.usermanagement.User;
+import de.focusshift.zeiterfassung.usermanagement.UserLocalId;
+import de.focusshift.zeiterfassung.usermanagement.UserManagementService;
 import jakarta.annotation.Nullable;
 import org.slf4j.Logger;
 import org.springframework.stereotype.Service;
