+++ src/main/java/de/focusshift/zeiterfassung/workingtime/WorkingTimeUpdateException.java
@@ -1,4 +1,4 @@
-package de.focusshift.zeiterfassung.usermanagement;
+package de.focusshift.zeiterfassung.workingtime;
 
 public class WorkingTimeUpdateException extends RuntimeException {
