+++ src/test/java/de/focusshift/zeiterfassung/workingtime/WorkingTimeCalendarTest.java
@@ -1,4 +1,4 @@
-package de.focusshift.zeiterfassung.usermanagement;
+package de.focusshift.zeiterfassung.workingtime;
 
 import de.focusshift.zeiterfassung.timeentry.PlannedWorkingHours;
 import org.junit.jupiter.api.Test;
